src/app/requirements.txt
deep_translator
Django
django-picklefield
ffmpeg
ffmpeg-python
google-cloud-texttospeech
googletrans
gTTS
librosa
moviepy
mysqlclient
nltk
openai-whisper
protobuf
pydub
pyffmpeg
scipy
six
social-auth-app-django
soundfile
transformers
TTS
ttsmms
torch